requirements.txt
python-telegram-bot==20.6
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
gspread>=5.12.0
oauth2client>=4.1.3
